Add unit test for empty arrays. (#45)
| Type | Description | Icon | Changelog |
| ---------- |
---------------------------------------------------------------------------|
:--: | :-------: |
| `test` | Making sure we don't break things later | 📌 | |

-------------------------------------------------------------------------------------------------------------

### 🛠️ Changes:
* added a unit test to verify that empty arrays are propagated through
the reactivity system when the first item is added.

### 📢 Additional Info:
This PR did not meet my hopes and dreams. I was hoping to write a unit
test that fails when 3899285aed7a81c4a5845686d63a71b58340ca83 is
reverted. But this test still passes. None the less I think it's a valid
unit test.

### 📚 Bookkeeping:

**Testing (if applicable)**:
- [X] Ran/wrote unit tests for this

**Checklist**
- [ ] Assigned PR to myself
- [ ] Added at least 1 person on the team as reviewer
- [NA] Release Notes: PRs types that have the :spiral_notepad: next to
them also require release notes to be added to the `CHANGELOG.md`