fix: make vue3 integration match vue2 (#48)
<!-- Remove all rows that don't apply to this PR --->
| Type | Description | Icon | Changelog |
| ---------- |
---------------------------------------------------------------------------|
:--: | :-------: |
| `fix` | Fixes an issue | 🐛 | ✅ |

-------------------------------------------------------------------------------------------------------------

### 🛠️ Changes:
<!-- Describe your changes in a brief, bullet list --->
- match the vue 2 reactivity implementation

### 📢 Additional Info:
<!-- Give some more context... Why is this needed? --->
Found some edge cases with the vue 3 reactivity implementation that
caused computed properties to "get stuck" and not update. The vue 2
reactivity implementation has worked very well in the past, so this
copies that. I'm sure there's a way to have more code sharing, but in
the interest of time, this will do for now.

### 📚 Bookkeeping:

**Testing (if applicable)**:
- [x] Ran/wrote unit tests for this

**Checklist**
- [x] Assigned PR to myself
- [x] Added at least 1 person on the team as reviewer
- [ ] Release Notes: PRs types that have the :spiral_notepad: next to
them also require release notes to be added to the `CHANGELOG.md`